Implement Authentication and Authorization in notes app

+++ src/api/notes/handler.js
@@ -16,9 +16,11 @@ class NotesHandler {
   async postNoteHandler(request, h) {
     try {
       this._validator.validateNotePayload(request.payload);
-      const { title = 'untitled', tags, body } = request.payload;
 
-      const noteId = await this._service.addNote({ title, body, tags });
+      const { title = 'untitled', tags, body } = request.payload;
+      const { id: credentialId } = request.auth.credentials;
+
+      const noteId = await this._service.addNote({ title, body, tags, owner: credentialId });
 
       const response = h.response({
         status: 'success',
@@ -50,7 +52,10 @@ class NotesHandler {
   }
 
   async getNotesHandler(request, h) {
-    const notes = await this._service.getNotes();
+    const { id: credentialId } = request.auth.credentials;
+
+    const notes = await this._service.getNotes(credentialId);
+
     const response = h.response({
       status: 'success',
       data: {
@@ -64,7 +69,11 @@ class NotesHandler {
   async getNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId);
       const note = await this._service.getNoteById(id);
+
       const response = h.response({
         status: 'success',
         data: {
@@ -94,8 +103,13 @@ class NotesHandler {
   async putNoteByIdHandler(request, h) {
     try {
       this._validator.validateNotePayload(request.payload);
+
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId)
       await this._service.editNoteById(id, request.payload);
+
       const response = h.response({
         status: 'success',
         message: 'Catatan berhasil diperbarui',
@@ -123,7 +137,11 @@ class NotesHandler {
   async deleteNoteByIdHandler(request, h) {
     try {
       const { id } = request.params;
+      const { id: credentialId } = request.auth.credentials;
+
+      await this._service.verifyNoteOwner(id, credentialId)
       await this._service.deleteNoteById(id);
+
       const response = h.response({
         status: 'success',
         message: 'Catatan berhasil dihapus',
